# movie-web-app/server/routers/movieApp-API-docs.json
{
    "openapi": "3.0.3",
    "info": {
      "title": "MovieApp API documentation",
      "version": "1.0"
    },
    "servers": [
      {
        "url": "http://localhost:3001"
      }
    ],
    "paths": {
...
      "/favorites": {
        "get": {
          "summary": "Get favorite movies",
          "description": "Retrieves the list of favorite movies for the authenticated user.",
          "responses": {
            "200": {
              "description": "A list of favorite movies.",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "favorites": {
                        "type": "array",
                        "items": {
                          "type": "string",
                          "example": "tt1234567"
                        }
                      }
                    }
                  }
                }
              }
            },
            "404": {
              "description": "User not found."
            }
          }
        },
        "post": {
          "summary": "Add a movie to favorites",
          "description": "Adds a movie to the authenticated user's list of favorites by movie ID.",
          "parameters": [
            {
              "name": "movieId",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "example": "tt1234567"
              },
              "description": "The IMDb ID of the movie to add."
            }
          ],
          "responses": {
            "201": {
              "description": "Movie added to favorites.",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "message": {
                        "type": "string",
                        "example": "Movie added to favorites"
                      }
                    }
                  }
                }
              }
            },
            "409": {
              "description": "Movie already in favorites."
            },
            "400": {
              "description": "Invalid movie ID."
            }
          }
        },
        "delete": {
          "summary": "Remove a movie from favorites",
          "description": "Removes a movie from the authenticated user's list of favorites by movie ID.",
          "parameters": [
            {
              "name": "movieId",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "example": "tt1234567"
              },
              "description": "The IMDb ID of the movie to remove."
            }
          ],
          "responses": {
            "200": {
              "description": "Movie removed from favorites.",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "message": {
                        "type": "string",
                        "example": "Movie removed from favorites"
                      }
                    }
                  }
                }
              }
            },
            "404": {
              "description": "Movie not found in favorites."
            }
          }
        }
      },
...
    }
  }